@@ -1,10 +1,6 @@
 #pragma once
 
-#include "GLHelpersBoost.h"
-
-#ifdef HAVE_DEVIL
 #include "GLHelpersDevil.h"
-#endif
 
 ////////////////////////////////////////////////////////////////////////////
 // Interface
@@ -38,11 +34,8 @@ inline GLuint LoadGLTextureFromFile(const char* path, size_t length)
     std::string extension;
     extension = filename.substr( filename.rfind( "." ) + 1 );
 
-    glTexId = LoadGLTextureUsingGIL(filename, extension.c_str());
 
-#ifdef HAVE_DEVIL
-    if(!glTexId) glTexId = LoadGLTextureFromDevIL(filename);
-#endif // HAVE_DEVIL
+    glTexId = LoadGLTextureFromDevIL(filename);
 
     return glTexId;
 }
@@ -52,10 +45,7 @@ inline GLuint LoadGLTextureFromArray(const unsigned char* data, size_t bytes, co
 {
     GLuint glTexId = 0;
 
-    if(!glTexId) glTexId = LoadGLTextureUsingGIL(data,bytes,extensionHint);
-#ifdef HAVE_DEVIL
-    if(!glTexId) glTexId = LoadGLTextureFromDevIL(data,bytes,extensionHint);
-#endif
+    glTexId = LoadGLTextureFromDevIL(data,bytes,extensionHint);
 
     return glTexId;
 }
