@@ -1,6 +1,6 @@
 #pragma once
 
-#include <SceneGraph/config.h>
+#include <SceneGraph/SceneGraphConfig.h>
 #include <SceneGraph/GLinclude.h>
 
 // OpenGL defines are in here
